requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.4
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.24.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
regex==2026.2.28
requests==2.32.5
sniffio==1.3.1
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
